Update dynamic position showcase. Probably needs to be changed since the introduction of KupManager.
