+++ packages/ketchup-showcase/src/plugins/router.ts
@@ -245,10 +245,10 @@ let javascriptRoutes = [
       import(`@/views/javascript/scrollonhover/Scrollonhover.vue`),
   },
   {
-    path: `/tippositioning`,
-    name: 'tippositioning',
+    path: `/dynamicposition`,
+    name: 'dynamicposition',
     component: () =>
-      import(`@/views/javascript/tippositioning/Tippositioning.vue`),
+      import(`@/views/javascript/dynamicposition/DynamicPosition.vue`),
   },
 ];
